+++ ntex-io/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## [0.3.0-beta.3] - 2023-06-21
 
-* Use .static_call() for dispatcher
+* Use static ContainerCall for dispatcher
 
 ## [0.3.0-beta.0] - 2023-06-16